feat: rename service to generator, use preset identifier from post request
